Switch to bitnami postgresql image
Changes
----
* Update: postgresql image to bitnami image
* Update: container security group
   * run as non-root
   * run with UID 1001
   * disable all linux capabilities
* Update: mounted filesystem group to non-root

Notes
----
The container can now be run as non-root and has had other restrictions applied to improve
the security.  However, the non-root user still belongs to the root group by default, as the
latter isn't set.  If `runAsGroup` is changed from the root GID, the container has insufficient
permissions to mount the PostgreSQL data volume.